refactor(meizam-api/axios-wrapper.ts):moved base url to axios wrapper
