+++ src/services/meizam-api/axios-wrapper.ts
@@ -1,10 +1,11 @@
 import axios, { AxiosRequestConfig } from 'axios';
-import { store } from '../store'
-import { notLoggedInException } from '../store/user/actions';
+import { store } from '../../store'
+import { notLoggedInException } from '../../store/user/actions';
 
-const baseUrl = process.env.REACT_APP_MEIZAM_API_BASE_HOST;
-const config: AxiosRequestConfig = {};
-if (baseUrl && baseUrl.indexOf('localhost') >= 0) {
+const config: AxiosRequestConfig = {
+  baseURL: process.env.REACT_APP_MEIZAM_API_BASE_HOST
+};
+if (config.baseURL && config.baseURL.indexOf('localhost') >= 0) {
   console.log('running on localhost')
 } else {
   config.withCredentials = true;
