+++ src/services/meizam-api/meizam-api.ts
@@ -1,9 +1,9 @@
-import axiosWrapper  from '../axios-wrapper'
+import axiosWrapper  from './axios-wrapper'
 import { Prediction, TeamPrediction, StagePrediction, StagePredictions } from '../../store/predictions/types';
 
 export default class MeizamApi {
   static async FetchUserInfo(): Promise<any> {
-    const response = await MeizamApi.get('/User/Info', {})
+    const response = await axiosWrapper.get('/User/Info')
     if (response.data.status != 'ok') {
       console.log(response.data)
       throw new Error(response.data.error)
@@ -27,7 +27,7 @@ export default class MeizamApi {
   }
 
   static async FetchGroupTable(groupId: number): Promise<any> {
-    const response = await MeizamApi.get(`/Group/Table`, { groupId: groupId })
+    const response = await axiosWrapper.get('/Group/Table', { params: { groupId: groupId } })
     if (response.data.status != 'ok') {
       console.log(response.data)
       throw new Error(response.data.error)
@@ -48,7 +48,7 @@ export default class MeizamApi {
     }
   }
   static async FetchPrediction(predictionId: number): Promise<Prediction> {
-    const response = await MeizamApi.get(`/Group/Prediction`, { predictionId: predictionId });
+    const response = await axiosWrapper.get('/Group/Prediction', { params: { predictionId: predictionId } });
       if (response.data.status != 'ok') {
         console.log(response.data)
         throw new Error(response.data.error)
@@ -69,20 +69,6 @@ export default class MeizamApi {
         groupsStage: MeizamApi.adaptPredictionStage(data.GroupStage),
         knockoutStage: MeizamApi.adaptPredictionStage(data.KnockoutStage)
       };
-  }
-
-  private static post(path: string, body: any): Promise<any> {
-    const baseUrl = MeizamApi.getBaseAndSetCredentials()
-    return axiosWrapper.post(`${baseUrl}${path}`, body)
-  }
-
-  private static get(path: string, params: any): Promise<any> {
-    const baseUrl = MeizamApi.getBaseAndSetCredentials()
-    return axiosWrapper.get(`${baseUrl}${path}`, { params })
-  }
-
-  private static getBaseAndSetCredentials(): string | undefined {
-    return process.env.REACT_APP_MEIZAM_API_BASE_HOST;
   }
 
   private static adaptPredictionStage(stage: any): StagePredictions {
